:hammer: Fixed index conversion error
See https://github.com/Gjacquenot/test_clang/runs/7747088752?check_suite_focus=true

[615/908] Building CXX object grpc/CMakeFiles/grpc.dir/src/SurfaceElevationFromGRPC.cpp.o
FAILED: grpc/CMakeFiles/grpc.dir/src/SurfaceElevationFromGRPC.cpp.o
/usr/bin/clang++ -DGTEST_HAS_POSIX_RE=0 -I. -I/opt/share/code/grpc/inc -I/opt/share/code/ssc -I/opt/share/code/exceptions/inc -I/opt/share/code/external_data_structures/inc -I/opt/share/code/core/inc -I/opt/share/code/environment_models/inc -I/opt/share/code/hdb_interpolators/inc -I/opt/share/code/mesh/inc -isystem /opt/eigen -isystem /opt/boost/include -isystem /opt/grpc/include -isystem grpc -isystem /opt/share/code/yaml-cpp/src -isystem /opt/share/code/yaml-cpp/include -isystem /opt/share/code/yaml-cpp/include/yaml-cpp -Werror -pedantic -pedantic-errors -Wextra -Wall -Wunused-function -Wunused-label -Wunused-parameter -Wunused-value -Wunused-variable -fno-common -Wformat=2 -Winit-self -Wpacked -Wpointer-arith -Wmissing-declarations -Wmissing-format-attribute -Wsign-compare -Wstrict-aliasing=2 -Wundef -Wconversion -Wno-misleading-indentation -Woverloaded-virtual -Weffc++ -Wwrite-strings -Wfatal-errors -Wno-deprecated -Wvariadic-macros -fPIC -std=gnu++11 -O3 -DNDEBUG -MD -MT grpc/CMakeFiles/grpc.dir/src/SurfaceElevationFromGRPC.cpp.o -MF grpc/CMakeFiles/grpc.dir/src/SurfaceElevationFromGRPC.cpp.o.d -o grpc/CMakeFiles/grpc.dir/src/SurfaceElevationFromGRPC.cpp.o -c /opt/share/code/grpc/src/SurfaceElevationFromGRPC.cpp
3852
/opt/share/code/grpc/src/SurfaceElevationFromGRPC.cpp:343:62: fatal error: implicit conversion changes signedness: 'int' to 'size_t' (aka 'unsigned long') [-Wsign-conversion]
            ssc::kinematics::PointMatrix ret("NED", response.vx_size());
                                         ~~~        ~~~~~~~~~^~~~~~~~~
1 error generated.

+++ code/grpc/src/SurfaceElevationFromGRPC.cpp
@@ -340,7 +340,7 @@ class SurfaceElevationFromGRPC::Impl
             std::copy(response.vx().begin(), response.vx().end(), std::back_inserter(vx));
             std::copy(response.vy().begin(), response.vy().end(), std::back_inserter(vy));
             std::copy(response.vz().begin(), response.vz().end(), std::back_inserter(vz));
-            ssc::kinematics::PointMatrix ret("NED", response.vx_size());
+            ssc::kinematics::PointMatrix ret("NED", static_cast<size_t>(response.vx_size()));
             for (int i = 0 ; i < response.vx_size() ; ++i)
             {
                 ret.m(0, i) = response.vx(i);
